disable warnings

+++ qsim/Perturbation1DImpl.cpp
@@ -20,8 +20,8 @@ void QuPerturbation1DImpl::InitPerturbation1D(std::function<Complex(Real)> const
 
 	if (fMet == SolverMethod::BornSerise) {
 
-		const_cast<int&>(fOrder) = opts.GetInt("order", 0);
-		const_cast<int&>(fSplit) = opts.GetInt("split_n", 0);
+		const_cast<int&>(fOrder) = (int)opts.GetInt("order", 0);
+		const_cast<int&>(fSplit) = (int)opts.GetInt("split_n", 0);
 
 		fPerturbationOptions.Init(opts);
 		if (fPerturbationOptions.fPreconditional) {
